src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backcall=0.2.0=py_0
blas=1.0=mkl
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py36h06a4308_0
cudatoolkit=10.2.89=hfd86e86_1
decorator=4.4.2=py_0
freetype=2.10.4=h5ab3b9f_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
jedi=0.18.0=py36h06a4308_1
jpeg=9b=h024ee3a_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36_0
lcms2=2.12=h3be6417_0
libedit=3.1=heed3624_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.0=h9df7e31_2
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openssl=1.0.2u=h7b6447c_0
parso=0.8.0=py_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=8.2.0=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=py36_0
pygments=2.7.1=py_0
python=3.6.2=hca45abc_19
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pyzmq=19.0.2=py36he6710b0_1
readline=7.0=ha6073c6_4
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=py_0
sqlite=3.23.1=he433501_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py36_cu102
tornado=6.0.4=py36h7b6447c_1
traitlets=4.3.3=py36_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
